creates error handler and complete crud heros

+++ src/services/HeroService.js
@@ -1,4 +1,5 @@
 const HeroRepository = require('../repositories/HeroRepository');
+const { DataNotFoundException } = require('../exceptions/DataNotFoundException');
 
 module.exports = {
   async getAll() {
@@ -6,25 +7,43 @@ module.exports = {
   },
 
   async getById(id) {
-    return await HeroRepository.getById(id);
-  },
-
-  async store(data) {
-
-    data.created_at = new Date;
-    data.updated_at = new Date;
-
-    const heroId = await HeroRepository.store(data);
-    const hero = await HeroRepository.getById(heroId);
+    const hero = await HeroRepository.getById(id);
+    if (!hero) {
+      throw new DataNotFoundException('Hero not found');
+    }
 
     return hero;
   },
 
+  async store(data) {
+    data.created_at = new Date;
+    data.updated_at = new Date;
+
+    const heroId = await HeroRepository.store(data);
+
+    return await HeroRepository.getById(heroId);
+  },
+
   async update(id, data) {
-    return await HeroRepository.update(id, data);
+    const hero = await HeroRepository.getById(id);
+    if (!hero) {
+      throw new DataNotFoundException('Hero not found');
+    }
+
+    await HeroRepository.update(id, data);
+    return await HeroRepository.getById(id);
   },
 
   async delete(id) {
-    return await HeroRepository.delete(id);
+    const hero = await HeroRepository.getById(id);
+    if (!hero) {
+      throw new DataNotFoundException('Hero not found');
+    }
+
+    await HeroRepository.delete(id);
+
+    return {
+      message: `Hero ${id} deleted!`
+    };
   }
 };
